requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distro==1.9.0
filelock==3.16.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
numpy==2.1.1
openai==1.45.0
packaging==24.1
pandas==2.2.2
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.3
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
requests==2.32.3
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
tiktoken==0.7.0
tomlkit==0.13.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
